consumer_service/requirements.txt
kafka-python==2.0.2
pymongo==4.8.0
elasticsearch==8.15.1